Refactor AI model configurations and delete unused files
This commit includes a major refactoring of the configuration of the AI models. It involves the deletion of the VectorDbConfig.java and EmbeddedModelFactory.java files, which are no longer needed. New configuration files ModelConfig.java and AwsConfig.java have been added for the configuration of the Anthropic3Chat model and AWS services respectively. Also, methods have been altered in the VectorStoreConfig.java and CohereModelFactory.java files. An AiFacade class was also added to handle AI interactions.