Update mappers to ignore unmapped targets
All mappers have been updated to ignore unmapped targets and to exclude the "id" field in the mapping process. This is achieved by setting the unmappedTargetPolicy to ReportingPolicy.IGNORE and by adding a Mapping annotation to ignore the "id" target. These changes will help in avoiding mapping errors when converting between domain entities and JPA entities.