Add service annotations and enhance storage service implementations
The commit introduces @Service and @Transactional annotations to make the classes ProjectStorageServiceImpl and EmployeeStorageServiceImpl spring managed beans. It also refactors the EmployeeStorageServiceImpl to bring in actual implementation for fetching available employees from the repository. It simplifies the injection of dependencies in the SuggestionService by leveraging @RequiredArgsConstructor annotation, removing the need for manual wiring and getters/setters for ProjectStorage and EmployeeStorage. Mock storage classes have been removed because of this enhancement.