Update EmployeeSuggester to return List
The EmployeeSuggester has been updated to return a List of Employees instead of an Optional. This change affects the `sagestEmployeesForProject` methods in AiPeopleManager and SuggestionService. A new file, TeamPrompts, has also been added to handle team suggestion prompts. The overall effect of this change is to simplify the code and make the operation of employee suggestions clearer.