Add method to explain AI suggestions for a project team
A new method explainSuggestion was introduced in multiple classes to provide a reason for the suggested team members for a given project. In addition to generating suggestions, the AI model can now also give a detailed explanation based on the project requirements and the list of employees. The AI runtime duration was also extended to facilitate this new feature.